moved the oled code from the keymap to the keyboard
