+++ keyboards/helix/rev3_4rows/rules.mk
@@ -6,3 +6,5 @@ OLED_DRIVER_ENABLE = yes
 ENCODER_ENABLE = yes
 DIP_SWITCH_ENABLE = yes
 LTO_ENABLE = yes
+
+SRC += oled_display.c
